Persist page when pressing save on select page screen
Fixes peristence of options by explicitly listing each property instead
of the whole object. The latter did not work because the object is a
class and exposed private fields prefixed with `_`.

Removes type guard usage on options config in the options repo, as it
seems rather fragile to rely on that guard to be maintained when
additional properties may be added in the future. By explicitly checking
for each property on the unknown object fetched from storage and
constructing the options config from scratch, we no longer rely on that
assumption.